feat(GrowingSetup): add GrowingSetup API and service implementation with status management

+++ src/main/java/dk/via/group1/urbanmicrofarm_backend/database/entities/GrowingSetupEntity.java
@@ -8,7 +8,7 @@ public class GrowingSetupEntity {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long setupId;
+  private int setupId;
 
   @Column(nullable = false)
   private String serialNumber;
@@ -18,8 +18,8 @@ public class GrowingSetupEntity {
   @Column(nullable = false)
   private String email;
 
-  public Long getSetupId() { return setupId; }
-  public void setSetupId(Long setupId) { this.setupId = setupId; }
+  public int getSetupId() { return setupId; }
+  public void setSetupId(int setupId) { this.setupId = setupId; }
 
   public String getSerialNumber() { return serialNumber; }
   public void setSerialNumber(String serialNumber) { this.serialNumber = serialNumber; }
